Using second autopairs plugin?
I'm tired of typing open parens in situations like this:
|word
()word
And then having to delete the closing parens.
nvim-autopairs can easily be configured to not do this, but I lost the endwise
in shell scripting. To avoid having to configure it now, I decided this weird
solution of having both of them.