Change some AwesomeWM key maps
Num pad enter now also opens the terminal

Switch increase/decrease client size so it makes more sense to me and resemble
more vim bindings. If "i o p ´" is "h j k l", then I think it makes sense to
have "i" (the "left", or "h") decrease master width factor, since most of the
time I am using the master and stack layout, so "i" makes sense. In the same
way, "´" as the "right", and thus increase. Also, since "o" would be the "j",
so "down", it makes sense to decrease height, and "p" to increase it, since it
would be the "k", so "up".