Add custom hightlight group autocmd tabs
In vim, the match command only works for a single window. This means that
after opening a new split, the custom group previously created with the
match would no longer exist. I had a custom group to change the tab color
to grey (default is blue), and when I create a new split, it would go back
to blue.