Fix togglebetweentestandfile script
There were two possible bugs:

The first one is that if the file was opened from somewhere not in the project's
root folder, the vim expand('%:h') would not have the expected string at the
start, and the variable "filename" would never be set.
Now insted of raising an exception, a message indicating tht the script only
works in the project's root is printed.
And to be honest, any folder higher than the project's root will work. If you
open the file from your home, it will work. It only has to find, in the elixir
example, a 'lib/' or 'test/' folder in the path. If you have another lib
directory inside other lib directory, it is going to think it is working, and it
might not work.
This bug was found by accident. The intended bug to be fixed was the second one:
The pattern match (lua's kinda of regex) was matching on '.ex', without scaping
the dot, making files named like "data_export.ex" being switched to a file named
"data_test.exsports_test.exs", which is clearly not intended.
The pattern was fixed to match on '%.ex$', which means that the dot has to be
literally a dot. As an extra, the extension has to be the last part of the file
name, making files named 'hey.ex.other' not match. I don't know if it is desired
or not, but it seemed better to do it like this.