[alacritty] Fix scrolling history and add font size
Before I was setting scrolling.history without
uncommenting the scrolling part first! So the
option was probably being ignored. Not a big deal,
though, since the default is probably good enough
and I use tmux most of the time, and not
alacritty's own scrolling tool.

Also, I added a font.size because on ubuntu (sad)
the initial font size was too big.