[awesomewm] Make my own screen temperature widget
The old one I was using had an internal state
controlling the temperature. I did not like the
approach very much, as it could unsync and was
also not possible, or not easy to, use in my
update_brightness_microphone_and_volume_icons
script.

The new one I made basically copying the screen
brightness widget, very similar to the volume and
microphone widgets as well. When updating, it
spawns its used command, in this case, xsct for
screen temperature.