[neovim] Big rewrite on nvim-tree and telescope keys functions to use lazy plugin spec to create the keymap directly
The `keys` function has to return a LazyKeys to
use with lazy.nvim. The returned value is a table
(list) where each item is what I'm gonna call a
"mapping". A mapping is as follows:

{
  [1] = left hand side,
  [2] = right hand side,
  mode = mode
}

So we can define what the mapping actually do from
lazy instead of calling vim.keymap.set from inside
the setup function.
But now if I require for example nvim-tree.api and
store it in a variable, like this:

  local nvim_tree_api = require('nvim-tree.api'),

This doesn't really work anymore. I think it is
because the plugin is not yet loaded when the
function `keys` is executed during the startup.
With a separate function that will just be
executed later, though, it is possible. This
separate function could just require the module
and return it. Silly, uh? All to avoid typing
require('nvim-tree.api') on each keymap...!