[neovim] Now use cokeline in next/prev buffer mappings
Using :bprev and :bnext would make vim follow the
order of the buffers according to their bufnrs,
which generally is very weird and not necessarly
the same as the order shown in the tabline.

ALso, it really does feel lonely not having a
:bprev/:bnext message on command line, but I
could't find a good thing to print.