[neovim] Move TabLine highlight from cokeline to catppuccin as TabLineFill
After a new cokeline update, it started using
TabLineFill by default to color the tab line fill
and it also made the choice of which highlight
group it uses configurable instead of hard coding
TabLine. Before I did not include the TabLine
color in catppuccin config because it looked weird
and kinda ugly on its own, so I figured it was a
setting I only wanted when using a buffer/tabline
plugin, like cokeline. But now I think it does
look good on its own, so I moved to the more
general place, the colorscheme config, which makes
a lot of sense in my opinion.
You can disable cokeline and any other tabline
plugin and then run :tabnew to see the tab fill
in action.