[neovim] Tweak and fix gitsigns mappings
Now don't centralize screen if there were no hunks
to move to when using next and previous hunk
mappings and also fix gitsigns attaching the
mappings to everybuffer on on_attach function,
instead of only current buffer. This could cause
issues if one buffer was not in a git dir and you
tried to use one of the mappings.