[neovim] Fix bug with Format on save by allowing Format to receive optional no_async argument
Since changing the auto command that formats
before write to use the user command Format, that
uses the standard vim lsp format with async set to
true, it would write the file to disc before
finishing formatting, thus modifying the file
again after the write, leaving the buffer in a
modified state and writing the unformatted file to
the disc.

Before this change, I used to call manually the
standard vim lsp format inside the callback of the
before write autocommand, and the async option was
set to false.

Now, Format accepts an argument (nargs = '?' means
that the function accepts 0 or 1 arguments), and
if it is exactly 'no_async', then the Format
command will call the lsp format with async set to
false.

And finally, the autocommand that triggers before
writing now calls 'Format no_async', solving this
issue and formatting synchronously, and then only
after finishing formatting, saving the buffer to
the disc.